Tactical fix until we can implement paging
On the GitHub teams GRaphQl call.

There are as of this commit now 106 teams in the DEFRA org and due to
the graphQl query only allowing a max of "first: 100" we are missing
some teams.

I have added a tactical fix of sorting by 'NAME" and getting the first
100 teams which currentl gies us teams A-T. But we obviously need to
impliment paging on this ASAP.