Creates missing deployments from lambda msg
This requires the deployment lambda to be updated to return the payload
it was triggerd with.
If we fail to find the corresponding CDP id for this message we can
assume it did originated from a different portal. Using the new request
data we can generate an almost complete deployment record for it and
save it.