CORE-945 fix for concurrent task update
Its possible two tasks are updated at the same time.
Since we read and replace the whole deployment model its possible to
loose updates.

This updates the handler to only set the specific fields and reload
them before working out the status.
We could put this all in a transaction but feels like an overkill for
now.

Also moves over to use entity data to work out if its a service/test
instead of the artifact.